final_thing/requirements.txt
branca==0.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
folium==0.16.0
idna==3.7
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.2
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
xyzservices==2024.4.0
